Move service user injection into the core
The injection of the service user can now be done by providing the
'inject_service_user' parameter for a listener. This injects fake user
credentials into the list of loaded users.

Currently, this parameter is not documented because parameters need to be
passed to the authentication modules when "initializing" them and the
parameter is embedded as a part of the SERV_LISTENER struct. A more
generic mechanism to pass parameters for the authenticators is needed.

A new entry point for the authenticators, which provides a mechanism
similar to the createInstance entry point of the routers and filters,
would allow a way to pass parameters for the authenticators.