Take new module declaration macro into use
Using the new module declaration macro gives a centralized place, where
all external points of entry are gathered, and gives us a strict set of
requirements for all modules.

The next step is writing a detailed document on the modules as a whole and
as individual module types. This document should contain the requirements
all modules must meet. It should also contain descriptions of the module
interfaces and how they are called by MaxScale.